backend/src/Final Backend/requirements.txt
flask
flask-cors
python-dotenv

llama-index
chromadb
sentence-transformers
transformers
torch
accelerate

# Optional but safe to include:
groq
